cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-bigquery-storage==2.*
google-cloud-secret-manager==2.*

